fix icon color
